+++ module.json
@@ -11,14 +11,13 @@
 			"path": "lang/en.json"
 		}
 	],
-    	"scripts": [
+    "scripts": [
 		"./scripts/settings/settings.js",
 		"./scripts/settings/settings-register.js",
-		"./scripts/gm-hot-seat-driver.js",
 		"foundry-hot-seat-observer.js" 
 	],
-    	"url": "https://github.com/zzarek/FoundryHotSeatObserver",
-    	"manifest": "https://raw.githubusercontent.com/zzarek/FoundryHotSeatObserver/HiddenTokens/module.json",
-    	"download": "https://github.com/zzarek/FoundryHotSeatObserver/archive/HiddenTokens.zip",
-	"minimumCoreVersion": "0.4.5"
+    "url": "https://github.com/zzarek/FoundryHotSeatObserver",
+    "manifest": "https://raw.githubusercontent.com/zzarek/FoundryHotSeatObserver/HiddenTokens/module.json",
+    "download": "https://github.com/zzarek/FoundryHotSeatObserver/archive/HiddenTokens.zip",
+	"minimumCoreVersion": "0.5.5"
 }
